Update Shop_Vac_Control.ino
Added a better theory of operation to comments.
Added a pinmode to define A0 as an input.
Identified that standalone operation was fixed by moving 5v input from VIN to +5v. The problem is that VIN is preregulator and when the USB cable is unplugged the power switches to the internal regulator. The output of the regulator is 4V when the input to it is 5v. Moving to the 5V pin is post regulator and insures the processor has 5v power